+++ src/librptexture/decoder/ImageDecoder_common.hpp
@@ -39,7 +39,7 @@
 #  define IMAGEDECODER_ALWAYS_HAS_SSE2 1
 #endif
 #ifdef HAVE_ARM_NEON_H
-#  ifdef RP_CPU_ARM64
+#  if defined(_WIN32) || defined(RP_CPU_ARM64)
 #    define IMAGEDECODER_ALWAYS_HAS_NEON 1
 #  endif
 #endif